Migrated all moderator commands
